Fix indexing loop when too many DWFs are reparsed
When there are many DWFs we end up in an indexation loop.

The manifest being merged triggers a reparse of DWFs. When there are 20+
files reparsed, it triggers a re-index which in turn triggers the
manifest from being re-merged, which then triggers a new reparse, etc.

This CL adds a key to identify which DWFs have a schema loaded for them.
We only reparse those files.

Fixes: 464216127
Test: existing + manually verified on a project with 20+ DWFs
Change-Id: I06d3effef32bd212950909b063ffe8c52cb13156

GitOrigin-RevId: dd4a23cea2b51621969dc9c091e800c5e96d2eb3

Space-RevId: 9b8aec18dbbf35a40e59ca50788ec9cc01f4f4f3
(cherry picked from commit ed4bef94cd6e7e3782b70bdf189615987ad3fb97)