Bring task tab to front on re-import
When re-importing an event-based artifact (.asdb) that is already
selected and open in a task tab, bring that tab to the foreground
instead of just re-selecting the session.

This makes the import behavior consistent with other artifact types and
improves the user experience.

Bug: 467227529
Test: Manual
Change-Id: Ib943f77013bb97f94846a88ef7d708eae94b076c

GitOrigin-RevId: 809efe69c4093bfd9bd085672bf507cc901da12f

Space-RevId: 04fe921aeed7a420424b4c5eab966ddae237e58f
(cherry picked from commit 9a40530352c6989d26204bf73c0667ca63c8b525)