Delay resolving delegate action
to avoid early action instantiations that may crash.

This change introduces a workaround. When replacing actions
conditionally we request the original action at the time of replacement
and if we receive a stubbed action, we reregister it under a different
id so that we can still unstub it later even after we replace the action
associated with the main id.

Bug: 464737975
Test: manually
Change-Id: I46422c8c50f5dcad70cbd2174283e2ceabc61edc

GitOrigin-RevId: 03ddf989295c83624aa2f291b1e8956c62aaadc9

Space-RevId: 90343471766b610ab0dcff389aa51d3784f15086
(cherry picked from commit fcc46368443012607763caaf5bf27cfbebc4a4fe)