Always request all languages
It is not expensive to request an empty output group. We should trust
the aspect and not the query data if they disagree.

This is primarily to reduce the scope of the BuildGraphData and prepare
it for dynamic loading.

Bug: n/a
Test: n/a
Change-Id: Iabcf55914628bcac136c7e4af6a9c164e6e69597

GitOrigin-RevId: ddd599fa6a5d9382fbdf0850aac0bc6afcaad587

Space-RevId: 7c4f695e4dcb5b1e0366fceebbac3de835e5e95d
(cherry picked from commit 2a9f7bc906d24d7a3912b99e8a8fbcdf3a1b0b8f)