Disable the action to build Run Configs for non-android Projects
As per the initial design of this new BuildAction, we had agreed to
build only Android RCs to begin with.
It would be a nice addition to be able to build the Run Configs for JVM
projects too since we already provide other build actions (project,
module, test scopes for example), but we can do this (if we decide to)
in a following change.

Bug: 467659028
Test: Added
Change-Id: Idbc7fdf0de0d38a4a0d9ec20e5acdd2df046b5d4

GitOrigin-RevId: d83a0bdd597bc4c34fcc99ad98c1962daf2a4878

Space-RevId: db2efa4b23a82ffa89c35b111536531294367582
(cherry picked from commit e4e13271b96fc7f6f56b4a4dbdcee383503331ce)