Override Kotlin version to latest for Compose templates
The Compose Compiler Gradle plugin is only compatible with Kotlin
versions greater than or equal to version 2.0.0. So when applying a
Compose-based template to a Kotlin project, we override the Kotlin
version if it's not compatible.

This change ensures the override is only applied if the version is not
compatible (version < 2.0.0), and uses the
`DEFAULT_KOTLIN_VERSION_FOR_NEW_PROJECTS` instead of a separate constant
to make it simpler to update.

Test: Added RenderTemplateModelTest
Bug: 322863175
Change-Id: I76257111af2abfc0216272be8b19c18bbae4eec7

GitOrigin-RevId: 7eb9de405ad4ced704bca29e043bd175a92ae9cf
(cherry picked from commit 533d03e5ddf4c49badf17b0ab9df80ac3d3c6429)