Revert "RAG: Enable new RAG Lucene indexer"
This reverts commit d5480640d9c401a6b3f7e9497eb7fc4b913baa41.

We experience the twice of indexing time because of LucIndex that is a
new version of RAG indexer. The main issue is the indexing and writing
the Lucene docs are done by only a single thread. We are working on
adding multiple threads. This revert is a temporary workaround to avoid
the twice of indexing time until we prepare the right solution.

(revert change-Id: I940b7b4ee2fe3bbc3344f245bea0682db6ee4278)

Bug: 462665363
Test: N/A
Change-Id: I030feada5fcb1fec69f484dbbe1c50aaa21de49e

GitOrigin-RevId: f6eec21516f8ad6f81d62f7a9024c5e975d1307b

Space-RevId: a4bf9b2d7cf871f1c065302f15bc7200105a7288
(cherry picked from commit 08efd1c654be6cb9549ee690f9e0211d55945756)