Fix Module used in AndroidAssetStudioAction
This ensures that we do not use the holder module since this module is
passed down to all the stack, including the New Android Vector Drawable
dialog. Because of that initial obtained module is used for rendering,
it can not be the holder module.

Test: AndroidAssetStudioActionTest
Fixes: 469339890
Change-Id: I6347c65428825a44dd68c01b6dab90d17b898cab

GitOrigin-RevId: f41b00049754e677b8b9137d25cdfbe0ee998bbd

Space-RevId: 1b97fd9a065e58826434e6bb19c8e2d44b675a48
(cherry picked from commit 11d77b6a1a5fd31f17d28d7351b4326b75db2e1a)