Do not generate BCP 47 LocaleQualifiers
The StringResourceData.localeSet are not using BCP 47 locales.
As a result the locales generated from LocaleManager.getLanguageCodes
should not be using BCP 47 either.

If the same approach is not used the usedLocales are not correctly
removed from the list of known locales in getUsedLocales.

Fixes: 450192746
Test: AddLocaleActionTest
Change-Id: I7508655c05e2678e6b9ea393a3c9e56a41e1da42

GitOrigin-RevId: e2a747004d011f32185fd4728d0dbe665051a5fa

Space-RevId: bebf57b162d25d5a5aa3a0190f69fe471962f2be
(cherry picked from commit f50971743570ec1a846471572583e44eeb1f4f19)